# apiDoc/api-docs.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /menus:
    put:
      tags:
      - Menu
      summary: "Edit dish name, dish price of menu"
      operationId: updateMenu
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditMenu'
        required: true
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /menus/findByPrice:
    get:
      tags:
      - Menu
      summary: "List all dishes that are within a price range, sorted by price or\
        \ alphabetically"
      operationId: findMenusWithinPrices
      parameters:
      - name: maxPrice
        in: query
        description: Max dish price (min = 1.0)
        required: true
        schema:
          minimum: 1
          type: number
      - name: minPrice
        in: query
        description: Min dish price
        required: false
        schema:
          type: number
          default: 0
      - name: sortByPrice
        in: query
        description: Sorted by price or dish name
        required: false
        schema:
          type: boolean
          default: true
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Menu'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /menus/findByDishName:
    get:
      tags:
      - Menu
      summary: Search dishes by dish name
      operationId: findMenusByDishName
      parameters:
      - name: dishName
        in: query
        description: The specified dish name
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Menu'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /open-hours:
    get:
      tags:
      - OpenHours
      summary: Find open hours by pagination
      operationId: findOpenHours
      parameters:
      - name: page
        in: query
        description: "The specified page, start from `0`"
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: The number of records in the page
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenHours'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /open-hours/findByTime:
    get:
      tags:
      - OpenHours
      summary: Find open hours by time
      operationId: findOpenHoursByTime
      parameters:
      - name: time
        in: query
        description: "The specified time (format is HH:mm), eg: 18:30"
        required: true
        schema:
          type: string
      - name: dayOfWeek
        in: query
        description: "A day of week (Sun = 0, Sat = 6)"
        required: false
        schema:
          maximum: 6
          minimum: 0
          type: integer
          format: int32
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenHours'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /open-hours/findByRestaurant:
    get:
      tags:
      - OpenHours
      summary: Find open hours by restaurant ID
      operationId: findOpenHoursByRestaurant
      parameters:
      - name: restaurantId
        in: query
        description: The specified restaurant ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenHours'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
...
components:
  schemas:
...
    AppError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        status:
          type: integer
          format: int32
        msg:
          type: string
...
    EditMenu:
      required:
      - dishName
      type: object
      properties:
        menuId:
          type: integer
          format: int64
        dishName:
          maxLength: 500
          minLength: 0
          type: string
        price:
          minimum: 0.0
          exclusiveMinimum: false
          type: number
    Menu:
      type: object
      properties:
        id:
          type: integer
          format: int64
        restaurantId:
          type: integer
          format: int64
        restaurantName:
          type: string
        dishName:
          type: string
        price:
          type: number
    OpenHours:
      type: object
      properties:
        id:
          type: integer
          format: int64
        restaurantId:
          type: integer
          format: int64
        restaurantName:
          type: string
        dayOfWeek:
          type: integer
          format: int32
        openPeriod:
          type: integer
          format: int32
        openHour:
          type: string
        closedHour:
          type: string
...